поправлена функция загрузки файлов на гуглдрайв. Увеличен таймаут

+++ backup_files_controller.py
@@ -15,18 +15,19 @@ async def backup_to_gdrive():
     if not result_dict:
         return
 
+    # Deleting remaining files (media)
+    if result_dict.get('files_list', None):
+        for file_ in result_dict['files_list']:
+            try:
+                os.remove(file_)
+            except Exception as e:
+                continue
+
     # Upload archives to Google Drive
     await google_drive.upload_files(result_dict['zip_file_list'])
     await google_drive_db.upload_files(['data.db'])
     # Delete zip files
     for file_ in result_dict['zip_file_list']:
-        try:
-            os.remove(file_)
-        except Exception as e:
-            continue
-
-    # Deleting remaining files (media)
-    for file_ in result_dict['files_list']:
         try:
             os.remove(file_)
         except Exception as e:
